Spy::Registry::Store => Spy::RegistryStore
